Authenticated refactor and test fixture creation:
+   Authenticated: creates authenticated defaults to decouple resourceModule and authenticated routes from the Authenticated component
+   Updates unit tests to use the defaults file in an authenticated fixture file
