+++ tests/unit/fixtures/admin/index.js
@@ -1,8 +1,9 @@
 import AppLayout from '@components/Ui'
 import Core from '@components/Core'
 import Unauthorized from '@components/Unauthorized'
-import defaults from '@components/Admin/src/defaults'
-import resourceModule from '@store/modules/resource'
+import defaults, {
+  authenticatedDefaults
+} from '@components/Admin/src/defaults'
 import { DefaultSidebar } from '@components/UiComponents';
 
 export default {
@@ -10,12 +11,6 @@ export default {
     ...defaults().props
   }
 }
-
-const unauthorizedRoutes = [{
-  path: '/unauthorized',
-  name: 'unauthorized',
-  component: Unauthorized
-}]
 
 export const Authenticated = {
   props: {
@@ -26,7 +21,6 @@ export const Authenticated = {
   },
   args: {
     Core,
-    resourceModule,
-    unauthorizedRoutes
+    ...authenticatedDefaults.args
   }
 }
